Create abstract board
